refactor(web): move sessions table into features/sessions
Preserve history via structure:move for SessionsTable and its session-row-data helpers.
