+++ web/src/features/sessions/session-row-data.clienttest.ts
@@ -1,6 +1,6 @@
 // @vitest-environment node
 
-import { joinSessionCoreAndMetrics } from "@/src/components/table/use-cases/session-row-data";
+import { joinSessionCoreAndMetrics } from "@/src/features/sessions/session-row-data";
 
 describe("joinSessionCoreAndMetrics", () => {
   it("keeps filtered core session fields when metrics contain all-time values", () => {
