+++ web/src/features/sessions/session-row-data.ts
