+++ web/src/features/users/UserDetailPage.tsx
@@ -9,7 +9,7 @@ import ScoresTable from "@/src/components/table/use-cases/scores";
 import { compactNumberFormatter, usdFormatter } from "@/src/utils/numbers";
 import { StringParam, useQueryParam, withDefault } from "use-query-params";
 import { DetailPageNav } from "@/src/features/navigate-detail-pages/DetailPageNav";
-import SessionsTable from "@/src/components/table/use-cases/sessions";
+import SessionsTable from "@/src/features/sessions/SessionsTable";
 import { cn } from "@/src/utils/tailwind";
 import { Badge } from "@/src/components/ui/badge";
 import { ActionButton } from "@/src/components/ActionButton";
